Add models.py and views.py files
