+++ dev_help.md
@@ -8,6 +8,12 @@ docker-compose up -d --build
 We can now view our flask app by visiting `http://localhost:5000/` in the browser.
 
 
+You can shut down the containers with:
+``` Bash 
+docker-compose down
+```
+
+
 If you are having trouble running locally, you may need to update the permissions to the `entrypoint.sh` file:
 ``` Bash
 chmod +x services/web/entrypoint.sh
@@ -17,3 +23,9 @@ To access the DB directly, execute:
 ``` Bash
 docker-compose exec db psql --username=dev --dbname=web
 ```
+
+You can view logs from inside the containers by executing:
+``` Bash
+docker-compose logs -f
+```
+To close the logs hit 'ctrl + c' twice.
