+++ services/web/project/__init__.py
@@ -1,4 +1,8 @@
-from flask import Flask, jsonify
+"""
+This file is used to instantiate the Flask app.
+"""
+
+from flask import Flask
 from flask_sqlalchemy import SQLAlchemy
 
 
@@ -6,18 +10,6 @@ app = Flask(__name__)
 app.config.from_object("project.config.Config")
 db = SQLAlchemy(app)
 
-
-class User(db.Model):
-    __tablename__ = "users"
-
-    id = db.Column(db.Integer, primary_key=True)
-    email = db.Column(db.String(128), unique=True, nullable=False)
-    active = db.Column(db.Boolean(), default=True, nullable=False)
-
-    def __init__(self, email):
-        self.email = email
-
-
-@app.route("/")
-def hello_world():
-    return jsonify(hello="world")
+# while it is usually better to do all imports at the top of the file, importing here mid-file is a best practice for Flask
+from project import models  
+from project import views
