+++ services/web/project/models.py
@@ -1,0 +1,15 @@
+"""
+This file contains all the database models for the Flask app.
+"""
+from project import db
+
+
+class User(db.Model):
+    __tablename__ = "users"
+
+    id = db.Column(db.Integer, primary_key=True)
+    email = db.Column(db.String(128), unique=True, nullable=False)
+    active = db.Column(db.Boolean(), default=True, nullable=False)
+
+    def __init__(self, email):
+        self.email = email
