--- services/web/project/routes.py
@@ -1,3 +1,0 @@
-"""
-This file will contain all the routes for the Flask app.
-"""
