+++ services/web/project/views.py
@@ -1,0 +1,8 @@
+"""
+This file contains all the routes for the Flask app.
+"""
+from project import app
+
+@app.route('/')
+def index():
+    return 'Hello World!'
